+++ src/repositories/index.ts
@@ -1,2 +1,2 @@
 export * as userRepository from './userRepository';
-export * as credentialsRepository from './credentialsRepository';
+export * as credentialRepository from './credentialRepository';
